Refactor service defaults extensions into separate files
Split Extensions.cs into HealthCheckExtensions, ServiceExtensions, and TelemetryExtensions for improved modularity and maintainability. Each concern (health checks, service defaults, telemetry) now resides in its own file, clarifying responsibilities and simplifying future updates.
