--- src/Neba.ServiceDefaults/Extensions.cs
@@ -1,165 +1,0 @@
-using Azure.Monitor.OpenTelemetry.AspNetCore;
-
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Diagnostics.HealthChecks;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
-using Microsoft.Extensions.Logging;
-
-using Neba.ServiceDefaults;
-
-using OpenTelemetry;
-using OpenTelemetry.Metrics;
-using OpenTelemetry.Trace;
-
-#pragma warning disable IDE0130 // Namespace does not match folder structure
-#pragma warning disable S1144 // Unused private types or members should be removed
-#pragma warning disable S2325 // Private types or members should be static
-namespace Microsoft.Extensions.Hosting;
-
-/// <summary>
-/// Adds common Aspire services: service discovery, resilience, health checks, and OpenTelemetry.
-/// This project should be referenced by each service project in your solution.
-/// To learn more about using this project, see https://aka.ms/dotnet/aspire/service-defaults
-/// </summary>
-public static class ServiceExtensions
-{
-    private const string HealthEndpointPath = "/health";
-    private const string AlivenessEndpointPath = "/alive";
-
-    /// <summary>
-    /// Extension members for <see cref="IHostApplicationBuilder"/> implementations.
-    /// </summary>
-    extension<TBuilder>(TBuilder builder) where TBuilder : IHostApplicationBuilder
-    {
-        /// <summary>
-        /// Adds common Aspire services: service discovery, resilience, health checks, and OpenTelemetry.
-        /// </summary>
-        /// <returns>The builder for chaining.</returns>
-        public TBuilder AddServiceDefaults()
-        {
-            builder.ConfigureOpenTelemetry();
-
-            builder.AddDefaultHealthChecks();
-
-            builder.Services.AddServiceDiscovery();
-
-            builder.Services.ConfigureHttpClientDefaults(http =>
-            {
-                // Turn on resilience by default
-                http.AddStandardResilienceHandler();
-
-                // Turn on service discovery by default
-                http.AddServiceDiscovery();
-            });
-
-            return builder;
-        }
-
-        /// <summary>
-        /// Configures OpenTelemetry for the application, including logging, metrics, and tracing.
-        /// </summary>
-        /// <returns>The builder for chaining.</returns>
-        public TBuilder ConfigureOpenTelemetry()
-        {
-            builder.Logging.AddOpenTelemetry(logging =>
-            {
-                logging.IncludeFormattedMessage = true;
-                logging.IncludeScopes = true;
-            });
-
-            builder.Services.AddOpenTelemetry()
-                .WithMetrics(metrics =>
-                {
-                    metrics.AddAspNetCoreInstrumentation()
-                        .AddHttpClientInstrumentation()
-                        .AddRuntimeInstrumentation();
-                })
-                .WithTracing(tracing =>
-                {
-                    tracing.AddSource(builder.Environment.ApplicationName)
-                        .AddAspNetCoreInstrumentation(tracing =>
-                            // Exclude health check requests from tracing
-                            tracing.Filter = context =>
-                                !context.Request.Path.StartsWithSegments(HealthEndpointPath, StringComparison.OrdinalIgnoreCase)
-                                && !context.Request.Path.StartsWithSegments(AlivenessEndpointPath, StringComparison.OrdinalIgnoreCase)
-                        )
-                        // Uncomment the following line to enable gRPC instrumentation (requires the OpenTelemetry.Instrumentation.GrpcNetClient package)
-                        //.AddGrpcClientInstrumentation()
-                        .AddHttpClientInstrumentation();
-                });
-
-            builder.AddOpenTelemetryExporters();
-
-            return builder;
-        }
-
-        private TBuilder AddOpenTelemetryExporters()
-        {
-            var useOtlpExporter = !string.IsNullOrWhiteSpace(builder.Configuration["OTEL_EXPORTER_OTLP_ENDPOINT"]);
-
-            if (useOtlpExporter)
-            {
-                builder.Services.AddOpenTelemetry().UseOtlpExporter();
-            }
-
-            // Uncomment the following lines to enable the Azure Monitor exporter (requires the Azure.Monitor.OpenTelemetry.AspNetCore package)
-            if (!string.IsNullOrEmpty(builder.Configuration["APPLICATIONINSIGHTS_CONNECTION_STRING"]))
-            {
-               builder.Services.AddOpenTelemetry()
-                  .UseAzureMonitor();
-            }
-
-            return builder;
-        }
-
-        /// <summary>
-        /// Adds default health checks to the application.
-        /// </summary>
-        /// <returns>The builder for chaining.</returns>
-        public TBuilder AddDefaultHealthChecks()
-        {
-            builder.Services.AddHealthChecks()
-                // Add a default liveness check to ensure app is responsive
-                .AddCheck("self", () => HealthCheckResult.Healthy(), ["live"]);
-
-            return builder;
-        }
-    }
-
-    /// <summary>
-    /// Extension members for <see cref="WebApplication"/>.
-    /// </summary>
-    extension(WebApplication app)
-    {
-        /// <summary>
-        /// Maps default endpoints for health checks.
-        /// </summary>
-        /// <returns>The application for chaining.</returns>
-        public WebApplication MapDefaultEndpoints()
-        {
-            #pragma warning disable CA1062 // Validate arguments of public methods
-
-            // Adding health checks endpoints to applications in non-development environments has security implications.
-            // See https://aka.ms/dotnet/aspire/healthchecks for details before enabling these endpoints in non-development environments.
-            if (app.Environment.IsDevelopment())
-            {
-                // All health checks must pass for app to be considered ready to accept traffic after starting
-                app.MapHealthChecks(HealthEndpointPath, new HealthCheckOptions
-                {
-                    ResponseWriter = HealthCheckResponseWriter.Default()
-                });
-
-                // Only health checks tagged with the "live" tag must pass for app to be considered alive
-                app.MapHealthChecks(AlivenessEndpointPath, new HealthCheckOptions
-                {
-                    Predicate = r => r.Tags.Contains("live")
-                });
-            }
-
-            #pragma warning restore CA1062 // Validate arguments of public methods
-
-            return app;
-        }
-    }
-}
