+++ src/Neba.ServiceDefaults/HealthCheckExtensions.cs
@@ -1,0 +1,56 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Diagnostics.HealthChecks;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Hosting;
+
+namespace Neba.ServiceDefaults;
+
+#pragma warning disable S1144 // Unused private types or members should be removed
+#pragma warning disable S2325 // Private types or members should be static
+
+internal static class HealthCheckExtensions
+{
+    internal const string HealthEndpointPath = "/health";
+    internal const string AlivenessEndpointPath = "/alive";
+
+    extension<TBuilder>(TBuilder builder)
+        where TBuilder : IHostApplicationBuilder
+    {
+        /// <summary>
+        /// Adds default health checks to the application.
+        /// </summary>
+        /// <returns>The builder for chaining.</returns>
+        internal TBuilder AddDefaultHealthChecks()
+        {
+            builder.Services.AddHealthChecks()
+                // Add a default liveness check to ensure app is responsive
+                .AddCheck("self", () => HealthCheckResult.Healthy(), ["live"]);
+
+            return builder;
+        }
+    }
+
+    extension(WebApplication app)
+    {
+        internal void MapDefaultHealthChecks()
+        {
+            // Adding health checks endpoints to applications in non-development environments has security implications.
+            // See https://aka.ms/dotnet/aspire/healthchecks for details before enabling these endpoints in non-development environments.
+            if (app.Environment.IsDevelopment())
+            {
+                // All health checks must pass for app to be considered ready to accept traffic after starting
+                app.MapHealthChecks(HealthEndpointPath, new HealthCheckOptions
+                {
+                    ResponseWriter = HealthCheckResponseWriter.Default()
+                });
+
+                // Only health checks tagged with the "live" tag must pass for app to be considered alive
+                app.MapHealthChecks(AlivenessEndpointPath, new HealthCheckOptions
+                {
+                    Predicate = r => r.Tags.Contains("live")
+                });
+            }
+        }
+    }
+}
