+++ src/Neba.ServiceDefaults/TelemetryExtensions.cs
@@ -1,0 +1,82 @@
+using System.Diagnostics;
+
+using Azure.Monitor.OpenTelemetry.AspNetCore;
+
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
+using OpenTelemetry;
+using OpenTelemetry.Metrics;
+using OpenTelemetry.Trace;
+
+namespace Neba.ServiceDefaults;
+
+#pragma warning disable S1144 // Unused private types or members should be removed
+#pragma warning disable S2325 // Private types or members should be static
+
+/// <summary>
+/// Extension methods for configuring telemetry in the application.
+/// </summary>
+public static class TelemetryExtensions
+{
+    extension<TBuilder>(TBuilder builder)
+        where TBuilder : IHostApplicationBuilder
+    {
+        /// <summary>
+        /// Configures OpenTelemetry for the application, including logging, metrics, and tracing.
+        /// </summary>
+        /// <returns>The builder for chaining.</returns>
+        internal TBuilder ConfigureOpenTelemetry()
+        {
+            builder.Logging.AddOpenTelemetry(logging =>
+            {
+                logging.IncludeFormattedMessage = true;
+                logging.IncludeScopes = true;
+            });
+
+            builder.Services.AddOpenTelemetry()
+                .WithMetrics(metrics =>
+                {
+                    metrics.AddAspNetCoreInstrumentation()
+                        .AddHttpClientInstrumentation()
+                        .AddRuntimeInstrumentation();
+                })
+                .WithTracing(tracing =>
+                {
+                    tracing.AddSource(builder.Environment.ApplicationName)
+                        .AddAspNetCoreInstrumentation(tracing =>
+                            // Exclude health check requests from tracing
+                            tracing.Filter = context =>
+                                !context.Request.Path.StartsWithSegments(HealthCheckExtensions.HealthEndpointPath, StringComparison.OrdinalIgnoreCase)
+                                && !context.Request.Path.StartsWithSegments(HealthCheckExtensions.AlivenessEndpointPath, StringComparison.OrdinalIgnoreCase)
+                        )
+                        // Uncomment the following line to enable gRPC instrumentation (requires the OpenTelemetry.Instrumentation.GrpcNetClient package)
+                        //.AddGrpcClientInstrumentation()
+                        .AddHttpClientInstrumentation();
+                });
+
+            builder.AddOpenTelemetryExporters();
+
+            return builder;
+        }
+
+        private TBuilder AddOpenTelemetryExporters()
+        {
+            var useOtlpExporter = !string.IsNullOrWhiteSpace(builder.Configuration["OTEL_EXPORTER_OTLP_ENDPOINT"]);
+
+            if (useOtlpExporter)
+            {
+                builder.Services.AddOpenTelemetry().UseOtlpExporter();
+            }
+
+            // Uncomment the following lines to enable the Azure Monitor exporter (requires the Azure.Monitor.OpenTelemetry.AspNetCore package)
+            if (!string.IsNullOrEmpty(builder.Configuration["APPLICATIONINSIGHTS_CONNECTION_STRING"]))
+            {
+               builder.Services.AddOpenTelemetry()
+                  .UseAzureMonitor();
+            }
+
+            return builder;
+        }
+    }
+}
